be/transform/requirements.txt
fastai==1.0.61
numpy==1.17.2
pandas==1.1.2 
torch==1.6.0
torchvision===0.7.0
Pillow==8.0.1
